% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2.R
\docType{data}
\name{data2}
\alias{data2}
\title{example of a dataset we could demonstrate a very basic operation of our func
days an dates are included}
\format{
An object of class \code{data.frame} with 3 rows and 2 columns.
}
\usage{
data2
}
\description{
example of a dataset we could demonstrate a very basic operation of our func
days an dates are included
}
\keyword{datasets}
