% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_apply.R
\name{df_apply}
\alias{df_apply}
\title{Apply alteration to dataframe}
\usage{
df_apply(.data, .fun, .filter, ...)
}
\arguments{
\item{.data}{A dataframe to alter.}

\item{.fun}{A function that can be applied to a column of a data frame and return a column of a data frame.}

\item{.filter}{A function that can be applied to a columns of a data frame and returns.}

\item{...}{Any other additional parameters passed to the function.}
}
\value{
An altered dataframe.
}
\description{
Apply alteration to dataframe
}
\examples{
x <- data.frame(c(1:7))
y <- function(.int) {return(.int * 2)}
z <- function(c) {is.numeric(c)}
df_apply(x, y, z)
}
