% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presidents.R
\docType{data}
\name{presidents}
\alias{presidents}
\title{Calvin University's Presidents}
\format{
A data frame with 12 observations on 3 variables.
name              name of president
begin_year        year in which the president took office
born_in_michigan  whether or not they were born in Michigan
}
\usage{
presidents
}
\description{
This data frame is a collection of information on Calvin University's presidents.
}
\examples{
if(require(ggplot2)) {
  presidents |>
    ggplot(aes(y = begin_year, color = born_in_michigan)) +
    geom_boxplot()
}
}
\keyword{datasets}
