% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prezident.R
\docType{data}
\name{prezident}
\alias{prezident}
\title{My data frame}
\format{
An object of class \code{data.frame} with 12 rows and 3 columns.
}
\usage{
prezident
}
\description{
This is a data frame containing information about Calvin presidents
}
\details{
My data frame
}
\section{Fields}{

\describe{
\item{\code{column1}}{A character column containing some information.}

\item{\code{column2}}{A numeric column containing some values.}

\item{\code{column3}}{A logical column indicating whether something is true or false.}
}}

\examples{
data(prezident)
head(prezident)

}
\keyword{datasets}
