% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_apply.R
\name{df_apply}
\alias{df_apply}
\title{Apply alteration to dataframe}
\usage{
df_apply(.data, .fun, .filter, ...)
}
\arguments{
\item{.data}{A dataframe to alter.}

\item{.fun}{A function that can be applied to a column of a data frame and return a column of a data frame.}

\item{.filter}{A function that can be applied to a columns of a data frame and returns TRUE or FALSE.}

\item{...}{Optional arguments to .fun.}
}
\value{
An altered dataframe.
}
\description{
Apply alteration to dataframe
}
\examples{
funct <- function(.int, .sigfigs) {return(signif((.int * 2), .sigfigs))}
df_apply(presidents, funct, is.numeric, .sigfigs = 2)

}
