% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presidents.R
\docType{data}
\name{presidents}
\alias{presidents}
\title{An example data frame about Calvin University's presidents}
\format{
An object of class \code{data.frame} with 12 rows and 3 columns.
}
\usage{
presidents
}
\description{
An example data frame about Calvin University's presidents
}
\keyword{datasets}
